backend/cloud-functions/doc-processing/main.py
import functions_framework
import os
from google.cloud import documentai_v1 as documentai
from google.cloud import vision

# TODO: Set your Google Cloud project ID and location
PROJECT_ID = os.environ.get('GCP_PROJECT')
LOCATION = os.environ.get('FUNCTION_REGION', 'us-central1') # Or specific Document AI region

# TODO: Configure Document AI processor ID and Vision features as needed
DOC_AI_PROCESSOR_ID = 'your-doc-ai-processor-id' # Replace with your processor ID

@functions_framework.cloud_event
def process_document(cloud_event):
    """Cloud Function triggered by a Cloud Storage event (e.g., file upload).
       Processes the uploaded document using Document AI and/or Cloud Vision.
    """
    if not PROJECT_ID:
        print("Error: GCP_PROJECT environment variable not set.")
        return

    data = cloud_event.data
    bucket = data.get("bucket")
    name = data.get("name")

    if not bucket or not name:
        print("Error: Invalid Cloud Storage event data. Missing bucket or name.")
        return

    gcs_uri = f"gs://{bucket}/{name}"
    print(f"Processing document: {gcs_uri}")

    try:
        # TODO: Determine file type (e.g., PDF, JPG, PNG) to decide processing path
        file_extension = name.split('.')[-1].lower()

        if file_extension in ['pdf', 'tiff', 'gif']: # Files suitable for Document AI
            print(f"Using Document AI for {name}")
            # TODO: Implement Document AI processing logic
            # Example:
            # opts = documentai.ClientOptions(api_endpoint=f'{LOCATION}-documentai.googleapis.com')
            # client = documentai.DocumentProcessorServiceClient(client_options=opts)
            # gcs_document = documentai.GcsDocument(gcs_uri=gcs_uri, mime_type='application/pdf') # Adjust mime type
            # request = documentai.ProcessRequest(
            #     name=client.processor_path(PROJECT_ID, LOCATION, DOC_AI_PROCESSOR_ID),
            #     gcs_document=gcs_document
            # )
            # result = client.process_document(request=request)
            # extracted_text = result.document.text
            # print(f"Document AI Extracted Text (first 50 chars): {extracted_text[:50]}...")
            pass # Placeholder

        elif file_extension in ['jpg', 'jpeg', 'png', 'bmp', 'webp']: # Files suitable for Cloud Vision
            print(f"Using Cloud Vision for {name}")
            # TODO: Implement Cloud Vision processing logic (e.g., OCR, label detection)
            # Example:
            # client = vision.ImageAnnotatorClient()
            # image = vision.Image()
            # image.source.image_uri = gcs_uri
            # response = client.text_detection(image=image)
            # texts = response.text_annotations
            # if texts:
            #     print(f"Vision API Extracted Text (first 50 chars): {texts[0].description[:50]}...")
            pass # Placeholder

        else:
            print(f"Unsupported file type: {file_extension} for file {name}")
            return

        # TODO: Store extracted data (e.g., in Supabase, BigQuery)
        print(f"Successfully processed (placeholder) {gcs_uri}")

    except Exception as e:
        print(f"Error processing document {gcs_uri}: {e}")
        # Consider adding error reporting or retry logic 